fix(azureresourcemonitoringservice): made tag checking/updating safer
Made sure that tag check only proceeds if AzureServiceResolver is able to retrieve a valid service
for given resource type. This is done because currently the resource types in the switch statement
for VM and Bastion are not valid. Resource type for Public IP, Network Interface and OS Disk is also
lacking...

fix Issue at:
https://dev.azure.com/equinor-core-dev/Sepes/_boards/board/t/DevOps/Stories/?workitem=1859