fix(dataset): assign custom role for dataset resource group (#699)
* fix(dataset): assign custom role for dataset resource group
Previously, the reader role was assigned for owners and sponsor rep, on the resource group containing the datasets. This role did not grant permission to create/read/delete blobs. A custom role has been created instead. This fix ensures that role assignment is set instead. The worker must have the id set in config, key is
Closes #692