test: improve study read rbac coverage (#714)
* refactor(rbac): OperationPermissionUtil.cs is now a service

* test: OperationPermissionService

* fix(study): added multiple allowed permissions to wbs validation
This endpoint required Study_Create permission, but since this endpoint is also used for those with edit study permission, we needed to specify multiple allowed roles. This had not been implemented before. Also written unit tests for it.
Closes #713


* refactor!: commented out unsupported dataset endpoints
Since pre-approved datasets is not yet supported, and somewhat unclear specification, it does not make sense to write tests for it yet. Therefore, these are now commented out until we bring those features back in.
BREAKING CHANGE: Removed support for CRUD operations for pre-approved datasets

* test: DatasetFileService